migrated forc_test::decode_logs -> forc_util::decode_logs
