+++ forc-test/Cargo.toml
@@ -14,7 +14,6 @@ forc-pkg.workspace = true
 fuel-abi-types.workspace = true
 fuel-tx = { workspace = true, features = ["test-helpers"] }
 fuel-vm = { workspace = true, features = ["random", "test-helpers"] }
-fuels-core.workspace = true
 rand.workspace = true
 rayon.workspace = true
 sway-core.workspace = true
