+++ test/src/e2e_vm_tests/mod.rs
@@ -11,7 +11,7 @@ use colored::*;
 use core::fmt;
 use forc_pkg::manifest::{GenericManifestFile, ManifestFile};
 use forc_pkg::BuildProfile;
-use forc_test::decode_log_data;
+use forc_util::decode_log_data;
 use fuel_vm::fuel_tx;
 use fuel_vm::fuel_types::canonical::Input;
 use fuel_vm::prelude::*;
